fix(test): id bindeado como i32 contra columna BIGSERIAL en test de @encrypted -- v1.147.1
CI en rojo tras v1.147.0, por un bug de test -- no un bug real. El
nuevo test de @encrypted contra Postgres real leia la fila cruda con
WHERE id = $1 bindeando id as i32, pero la columna real es BIGSERIAL
(int8) -- el driver de Postgres rechaza el tipo. Corregido a bindear
id (i64) directo, sin cast. El cifrado en si (write_param/decode_row)
ya funcionaba: 76 de 77 tests pasaron en CI.

+++ compiler/tests/pg_integration.rs
@@ -4554,7 +4554,7 @@ fn an_encrypted_field_stores_ciphertext_in_postgres_but_round_trips_to_the_exact
     // descifrado.
     let mut admin = postgres::Client::connect(&url, postgres::NoTls).expect("conectar como admin");
     let raw: String = admin
-        .query_one(&format!("SELECT ssn FROM \"{COLLECTION}\" WHERE id = $1"), &[&(id as i32)])
+        .query_one(&format!("SELECT ssn FROM \"{COLLECTION}\" WHERE id = $1"), &[&id])
         .expect("leer la fila cruda")
         .get(0);
     assert_ne!(raw, "123-45-6789", "el ssn NUNCA debe estar en texto plano en la columna física: {raw:?}");
